requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas
ollama==0.1.6
python-dotenv==1.0.0
openai
numpy==2.2
scikit-learn
gunicorn